+++ app/README.md
@@ -11,4 +11,5 @@
 Firebase Crashlytics | - | 利用可能
 google-services.json | - | 必須
 Timber | 全て出力 | `e()` をCrashlytics へ転送
+カスタムApplication クラス | `DebugApplication` | `ReleaseApplication`
 署名情報(release.jks) | - | 必須
