+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/MainApplication.kt
@@ -1,26 +1,12 @@
 package jp.co.yumemi.android.code_check
 
 import android.app.Application
-import android.util.Log
-import com.google.firebase.crashlytics.ktx.crashlytics
-import com.google.firebase.ktx.Firebase
 import timber.log.Timber
 
-open class MainApplication : Application() {
+abstract class MainApplication : Application() {
 
     /** Timber のログ出力ツリー */
-    protected open val timberTree = object : Timber.Tree() {
-        override fun log(
-            priority: Int,
-            tag: String?,
-            message: String,
-            t: Throwable?,
-        ) {
-            when (priority) {
-                Log.ERROR -> t?.also { Firebase.crashlytics.recordException(it) }
-            }
-        }
-    }
+    protected abstract val timberTree: Timber.Tree
 
 
     override fun onCreate() {
